feat: v3.1.5 Update (主要内容为官机) / 扫雷小游戏调整为6 * 5棋盘，默认6雷 / 调整Webui中的部分渲染内容 + 将交互事件下分到按钮点击事件和私聊设置更改事件，用于进一步的适配 + 更新了私聊主动消息的相关内容，私聊推送尚未制作 / 调整了新闻总结的某些提示词内容 + 开放了一些公开查询API接口 + 更新了新的帮助菜单，制作了参考文档 / 修复了一些已知问题

+++ src/main/java/top/yzljc/atribot/auth/official/OfficialGroups.java
@@ -67,6 +67,7 @@ public class OfficialGroups {
      * 删除群数据
      */
     public static boolean removeGroup(String groupOpenId) {
+        if (isGroupBlacklisted(groupOpenId)) return false; // 黑名单群不允许删除
         if (GroupRepository.deleteGroup(groupOpenId)) {
             cache.remove(groupOpenId);
             return true;
